backend/requirements.txt
flask
python-dotenv
PyPDF2
langchain
faiss-cpu
langchain-community